Error Checks
- Check if a field is final before assigning
- Check if a method is final when being overridden
- Check if a class is final when being extended
- Check if the extended class is an actual class
- Check if the implemented interfaces are actual interfaces

+++ src/main/java/dyvil/tools/compiler/ast/api/IContext.java
@@ -5,11 +5,11 @@ import dyvil.tools.compiler.ast.method.MethodMatch;
 
 public interface IContext
 {
-	public static final int	INVISIBLE			= 0;
-	public static final int	READ_ACCESS			= 1;
-	public static final int	WRITE_ACCESS		= 2;
-	public static final int	READ_WRITE_ACCESS	= 3;
-	public static final int	STATIC				= 4;
+	public static final byte	INVISIBLE			= 0;
+	public static final byte	READ_ACCESS			= 1;
+	public static final byte	WRITE_ACCESS		= 2;
+	public static final byte	READ_WRITE_ACCESS	= 3;
+	public static final byte	STATIC				= 4;
 	
 	public boolean isStatic();
